Remove git client plugin test exclusions
Command line git 2.38.1 and backports to earlier versions now exclude
submodule clones from a file:// URL. The tests in the git client
plugin rely on submodule clones from a file:// URL for some of the
tests.  Git client plugin 2.13.0 includes test changes to allow the
tests to run on git 2.38.1 successfully.  These exclusions should not
be needed any longer.

See https://github.com/jenkinsci/bom/pull/1484 for the original addition
of the exclusions.

https://github.com/jenkinsci/git-client-plugin/pull/920#issuecomment-1297169331
reminds us to remove the exclusion.

https://github.com/jenkinsci/git-client-plugin/pull/923 implements the
fix in the tests to allow file:// protocol clones during the submodule
operations in the tests.