Pin git client plugin 3.12.4 for Jenkins 2.332.x tests (#1643)
https://issues.jenkins.io/browse/JENKINS-70174 notes that the git client
plugin automated tests failed when using a shallow clone (depth == 1)
of a single branch.  The issue was fixed for Jenkins 2.346.3 and newer
with the release of git client plugin 3.13.1.  This release is a backport
of that fix.

No plugin functionality was changed in this release.  It changes the
automated tests when run from a shallow clone (depth == 1) of a single
branch.

https://github.com/jenkinsci/bom/pull/1613 is the key consumer in the
plugin bill of materials.

https://github.com/jenkinsci/plugin-compat-tester/pull/382 is the key
consumer in the plugin compatibility tester.