Drop 2.361.x from plugin BOM
https://www.jenkins.io/security/advisory/2023-03-08/ notes a security
vulnerability in Jenkins versions prior to 2.375.4.  Let's drop support
for 2.361.x to reduce the testing load and to encourage users to upgrade
and avoid security issues.

98% of the 96000 installations of git plugin 5.0 were already running
2.375.1 or newer as of 1 May 2023.  Over 40% of all git plugin
installations were using 2.375.1 or newer as of 1 May 2023.

Users that are upgrading to within 6 months of the most recent release
(about 1/3 of the total installed base) are upgrading both core and
plugins.