Temporarily exclude ExtendedEmailPublisherDescriptorTest (#2167)
UI changes in https://github.com/jenkinsci/jenkins/pull/7717 changed
the contents of the DOM that the groovyClassPath() test was using for
its assertion.  Temporarily ignore that test result until a pull request
to email-ext resolves the issue.

The email-ext pull request is not ready but I will do my best to have
it ready within 48 hours so that this exclusion can truly be temporary.