Retry up to 6 times
https://ci.jenkins.io/job/Tools/job/bom/job/master/1795/ had 13 tasks
that reached 3 retries and at least one of those tasks failed even on
the third retry.  Let's allow up to 6 retries in hopes of having all 450+
tasks complete successfully within 6 retries.

Please don't merge this until the current build of the master branch
has completed.  If it is merged while the current build is running,
it will cancel the running build.