Exclude a git client maintenance test (#2209)
Fails for command line git 2.41 with git client plugin releases prior
to 4.4.0 because it assumes that only two files are written for a loose
object.  Command line git 2.41 writes a third file for a loose object.

Git client plugin 4.4.0 requires git 2.387.3 or later, so its correction
of the test does not resolve the issue for the 2.375.x BOM test.

Remove this exclusion when 2.375.x is no longer being tested.