Pin snakeyaml-api 1.33 for Jenkins 2.387.3
This pinning of the plugin version is being done to pass the 2.387.3
tests for the blue ocean plugin and the kubernetes plugin.  I think
that is better than excluding the failing tests, since the failing tests
would be excluded on all versions, not just one version.

This is atypical, since the snakeyaml 2.2 plugin supports Jenkins 2.387.3,
but other plugins that run on 2.387.3 do not support snakeyaml 2.2.

The blue ocean plugin version 1.27.5.1 (most recent release supported
with Jenkins 2.387.3) does not support snakeyaml plugin 2.2 based on
https://issues.jenkins.io/browse/JENKINS-71966

The kubernetes plugin version 4007.v633279962016 (most recent
release supported with Jenkins 2.387.3) does not seem to
support snakeyaml plugin 2.2 based on the test results from
https://github.com/jenkinsci/bom/pull/2464/checks?check_run_id=16489432297