Pin previous credentials-binding release for LTS profiles (#2514)
The most recent release of the credentials-binding plugin adds masking
for base64 credentials.  That's a nice improvement.  Unfortunately,
it causes one of the config-file-provider tests to fail.

Adapt older bom profiles to "Bump credentials-binding (#2509)" by
retaining the current version of the credentials binding plugin on
the weekly release and pinning the previous credentials binding plugin
release on the LTS releases.

Could exclude the test failure on all releases, but it seemed better
to be able to detect test failures from the weekly release even if we
can't yet test the new version with the LTS releases.

This partially reverts commit bab8257f69b4411d683a420b106d8361fae5d0cf.