Revert "Bump io.jenkins.plugins:gitlab-branch-source in /bom-weekly (#2548)" (#2552)
The 679.v1dfd3604d46e release of the GitLab branch source plugin adds
the symbol `gitLab` to its implementation of a repository browser in
PR-360. The repository browser implementation is used by the git plugin to
allow the user to choose the preferred way of displaying links to changes.

Unfortunately, the git plugin already defines the symbol `gitLab` for
a repository browser implementation. The addition of the new symbol
to the GitLab branch source plugin breaks the configuration as code
definitions of git plugin users if they have used the gitLab symbol in
their definition of global Pipeline libraries.

https://github.com/jenkinsci/gitlab-branch-source-plugin/pull/360#issuecomment-1741737248
has more details including screenshots.

https://github.com/jenkinsci/bom/runs/17272883656 detected the failure
while testing the plugin bill of materials with git plugin 5.2.0 and
GitLab branch source plugin 679.v1dfd3604d46e.

https://issues.jenkins.io/browse/JENKINS-72104 records the GitLab branch
source plugin bug that this is reverting.

This reverts commit d5e912b1d2df82d0c35cd6b82bcfa46737d86f3b.