Test the copy button logic improvement in core
HTMLUnit 3 should already be implemented in all the plugins that are
included in the bill of materials, but it is better to verify in the
plugin bill of materials now than to discover it with a broken build of
the plugin BOM at release time.

Checking

* https://github.com/jenkinsci/jenkins/pull/8554

Thanks to @D3SOX for the pull request proposing the improvement.