Exclude MultiBranchTest.multiBranchPipelineIndex test
Would be better to exclude only for Jenkins 2.387.x, since the test has
been fixed in https://github.com/jenkinsci/blueocean-plugin/pull/2505 ,
but there is not currently a way to exclude a test for a single line.

Includes a TODO comment that the exclusion should be removed when the
2.387.x line is no longer included in the plugin bill of materials.

https://github.com/jenkinsci/bom/pull/2593/checks?check_run_id=17920234324
shows the most recent failure only happeend on the 2.387.x line.