Revert "Exclude AggregatedTestResultPublisherTest"
https://github.com/jenkinsci/junit-plugin/releases/tag/1252.vfc2e5efa_294f
is the JUnit plugin release that fixes the test.

Special thanks to Tim Jacomb for fixing the test in
https://github.com/jenkinsci/junit-plugin/pull/591

https://github.com/jenkinsci/junit-plugin/issues/588 is the issue report
for the JUnit plugin tests that depend on the icon based build status
display.

https://github.com/jenkinsci/jenkins/pull/8705 is the Jenkins core
pull request that switched from using icon based build status to symbol
based build status.  That change also fixes a layout issue in the Safari
web browser.

This reverts commit cb2376ef4fe26b68fef13139290b5e65ddc447ec.

Author: Mark Waite <mark.earl.waite@gmail.com>
Date:   Tue Nov 28 22:00:27 2023 -0700