Block update to ssh-credentials plugin 326.v7fcb_a_ef6194b_
https://github.com/jenkinsci/ssh-credentials-plugin/pull/199 makes the
trilead-api plugin an optional dependency of the ssh-credentials plugin.
That exposed a latent issue in the subversion plugin that is resolved
in https://github.com/jenkinsci/subversion-plugin/pull/284 .

https://github.com/jenkinsci/bom/issues/3050 describes the details of
the test failures in the plugin bill of mateerials.  The failing tests
can be seen with the commands:

 PLUGINS=subversion TEST=CompareAgainstBaselineCallableTest bash local-test.sh
 PLUGINS=mina-sshd-api-core LINE=weekly bash local-test.sh

No need to block later releases, since the pull request thet fixes the
issue has been submitted and is ready for review.