Test parameterized trigger optional dependency update
The parameterized trigger plugin has an optional dependency on
promoted builds plugin 3.11.  The git plugin also declares an optional
dependency on the promoted builds plugin 3.11.  Attempts to update that
optional dependency in the git plugin have failed.  This is an attempt
to upgrade the dependency in the promoted builds plugin first, in hopes
that will eventually allow the dependency to be updated in the git plugin.

Tests the plugin built from:

* https://github.com/jenkinsci/parameterized-trigger-plugin/pull/378