Do not set the JTH War file
as mentioned in the upstream PR that attempted to do this
unconditionally this is broken by design.

the JTH and tests assume that the classpath is correctly setup however
it is not, as demonstrated in the prior commit.

According to Jenkins there is a plugin installed (trilead-api) because
it is in the megawar, however its classess are not in the flat classpath
used by JenkinsRule and thus breakage ensues.

ref: https://github.com/jenkinsci/plugin-compat-tester/pull/470#issuecomment-1448789132