Do not set the JTH War file
as mentioned in the upstream PR that attempted to do this
unconditionally this is broken by design.

the JTH and tests assume that the classpath is correctly setup however
it is not, as demonstrated in the prior commit.

According to Jenkins there is a plugin installed (trilead-api) because
it is in the megawar, however its classess are not in the flat classpath
used by JenkinsRule and thus breakage ensues.

There are also issues where a plugin is present that is incompatable
with a mode of running for the test - e.g. a system property is set when
running JTH to test a mode of running for a plugin (e.g. FIPS) and the
plugin itself is incompatable with that mode (which causes a
BootFailure).

ref: https://github.com/jenkinsci/plugin-compat-tester/pull/470#issuecomment-1448789132
fixes: #3231