Check the modernise build time trend page pull request
https://github.com/jenkinsci/jenkins/pull/9465 improves the build time
trend page by adding a "time since" column and an explicit icon that links
to the console output.  It also hides the agent column for Pipeline jobs
and allows the table to be resized.

Check that those very nice improvements do not break any of the tests
in the plugin bill of materials.