Add Groovy Postbuild plugin to managed set (#3751)
Fix https://github.com/jenkinsci/bom/issues/3635

Plugin has over 19000 installations and is used regularly when testing
new Jenkins releases.  Was affected by the badge 2.0 compatibility break
and has been adapted to that change in the release that is included
in bom-weekly/pom.xml.

Confirmed that tests pass for this plugin with:

PLUGINS=groovy-postbuild bash local-test.sh

Also confirmed that the tests fail with the older version of the plugin.