Further simplify scripts by removing issueId parameter
After a bit more digging, I found that it wasn't necessary to pass in the GitHub issue id to some of the scripts. This should make it even easier to follow the daily (primarily Friday) task list.

The catch is the math to get the issue id is we are looking for an issue that has a `release` label (set in the create script). We need to make sure that there is never more than one open issue that has a `release` label. I did limit the return to just 1, but if there are multiple, it's going to grab the most current, even if that's not the one you want.  This probably isn't a huge deal because only one person should be doing the release, so that's why I didn't go overboard with protections.