Simplify release manager scripts
After following the "day in the life" task list and actually using the scripts, there were quite a few things that were (a) wrong [even though tested], (b) missing and (c) annoying. This PR updates the task list and scripts to hopefully a little more straightforward.

Don't be surprised if we further simplify in the future.