Pin htmlpublisher 1.36 for older lines
Prevent test failure in the declarative Pipeline plugin when it attempts
to load htmlpublisher 1.37 on Jenkins 2.479.x and earlier.

Tests fail with the commands:

$ LINE=2.479.x PLUGINS=pipeline-model-definition TEST=InjectedTest bash local-test.sh
$ LINE=2.462.x PLUGINS=pipeline-model-definition TEST=InjectedTest bash local-test.sh
$ LINE=2.452.x PLUGINS=pipeline-model-definition TEST=InjectedTest bash local-test.sh

Failure message is:

> Failed to load: HTML Publisher plugin (htmlpublisher 1.37)
>  - Update required: Matrix Project Plugin (matrix-project
>    832.va_66e270d2946) to be updated to 838.v4d7b_7b_f9b_d4b_ or higher

The matrix project release 839.vff91cd7e3a_b_2 is already defined in those
older release lines but that definition does not seem to be honored by
the pipeline-model-definition test configuration.