Pre-heat the artifact caching proxy (#4095)
* Pre-heat the artifact caching proxy

Attempt to detect artifact caching proxy failures earlier in the build
process by downloading the artifacts during the prep stage.  The prep
stage may not need the dependencies, but it is less expensive if the
download fails early in the build process rather than failing later
during plugin compatibility testing.

This is flawed because it does not account for the number of artifact
caching proxy instances that may be running.  It might pre-heat only
one of the caches, but hopefully that is better than nothing.

* Use consistent form of --no-transfer-progress

Argument was already used elsewhere as `-ntp` so it should use the
same form here.