Test docker-workflow pre-release 596.v942f78663d94
The 595.v60cec912059b_ tests fail on the ci.jenkins.io agents that are
used for BOM builds.  The failure message is:

> java.lang.IllegalArgumentException: No enum constant
> org.jenkinsci.plugins.docker.workflow.DockerTestUtil.DockerOsMode.
> CANNOT CONNECT TO THE DOCKER DAEMON AT UNIX:///VAR/RUN/DOCKER.SOCK. IS
> THE DOCKER DAEMON RUNNING?

A pull request has been proposed to resolve the failure by skipping
those tests on machines that have the docker executable available but
the user does not have permission to use docker.  The pull request is:

* https://github.com/jenkinsci/docker-workflow-plugin/pull/336

This commit tests that build in plugin BOM