Drop the 2.462.x line (#4650)
The transition to the 2.479.x line is progressing very well.  That is
causing the 2.462.x line to become an independent, outdated copy
of bom-weekly.

There are 1800 lines in bom-weekly and 1500 lines in bom-2.462.x.
New releases of the bom-2.462.x line will have far fewer plugin updates
than is typical for the oldest line in the bom.

Removing the 2.462.x line may also allow us to successfully complete
a full build on ci.jenkins.io.  That is a temporary condition, but it
will allow us more time to identify the issues on ci.jenkins.io and
resolve them.