Test git client 6.1.3 incremental
Test in 6.1.2 fails with Apache Mina SSHD 2.15.0.

https://github.com/jenkinsci/git-client-plugin/pull/1274 improves the
tests on the master branch so that the plugin is ready for JGit 7.2.0
and its requirement for Apache Mina SSHD 2.15.0.

Test assertion is disabled on the 6.1.3 branch so that it will pass the
test with both Apache Mina SSHD 2.14.x and with Apache Mina SSHD 2.15.0.