Update release lead calendar
Move Bruno Verachten's assignment as release lead to be later so that
Bruno covers the BOM release instead of Basil on 4 Jul 2025.  That day
is a public holiday for Basil in the U.S.

Avoids the week of Establishment Day (1 Jul 2025) in Hong Kong for
Kris Stern.

Avoids Labour Day (1 May 2025), Victory in Europe Day (8 May 2025),
and Bastille Day (14 Jul 2025) in France for Bruno Verachten.

Avoids Memorial Day (26 May 2025) in the U.S. for Mark Waite.