Pin kubernetes plugins for 2.492.x due to kubernetes-credentials-provider
The kubernetes-credentials-provider had to be pinned for 2.492.x because
the latest release requires 2.504.1 or newer due to its dependency on
openstack-cloud plugin.

Pin the updated plugins so that the kubernetes-credentials-provider
tests pass.

Testing done:

Confirmed that this fails without the change:

PLUGINS=kubernetes-credentials-provider TEST=KubernetesCredentialsProviderTest#credentialScope LINE=2.492.x bash ./local-test.sh

and passes with the change

Initially detected while evaluating:

* https://github.com/jenkinsci/bom/pull/5243

Test failure visible in CI job:

* https://ci.jenkins.io/job/Tools/job/bom/job/PR-5243/3/testReport/