Move release instructions to a separate file
Daniel Beck noted that each time there is a modification to the
CONTRIBUTING.md file, GitHub shows an indicator that warns of a change
in the contributing guidelines.  Changing the release lead assignment
is not a change of the contributing guidelines.

Let's move the documentation related to releasing the BOM into a separate
file that is dedicated to the work of the release leads.  Contributors
that are not release leads will continue to read CONTRIBUTING.md.