Exclude BitbucketPipelineCreateRequestTest in Blue Ocean tests
Bitbucket branch source plugin intentionally changed the order of HTTP
requests, but the Blue Ocean tests are checking a precise match with
wiremock.  Allow either the old or the new initial URL to be matched.

Test fix submitted as pull request:

* https://github.com/jenkinsci/blueocean-plugin/pull/2654

Test fix uses a pattern match in Bitbucket Wiremock test definition so
that automated tests pass with previous and latest versions of Bitbucket
Branch Source plugin.  Special thanks to @nfalco79 for investigation
and the original submission in pull request:

* https://github.com/jenkinsci/blueocean-plugin/pull/2653

Bitbucket branch source 936.x and 937.x use different initial URLs.
Check for either and pass the test if one or the other is used.

Original change included in Bitbucket branch source 937.0.0 pull request:

* https://github.com/jenkinsci/bitbucket-branch-source-plugin/pull/1060

Originally detected in:

* https://github.com/jenkinsci/bom/pull/5669

Confirmed that the test fails without this exclusion.