Exclude ldap and azure-vm-agents tests of configuration as code (#5771)
Configuration as code plugin 1985.vdda_32d0c4ea_b_ includes the 'allow
empty map' fix from pull request:

* https://github.com/jenkinsci/configuration-as-code-plugin/pull/2712

That fix needed changes in the tests of 3 plugins:

* azure-vm-agents: AdvancedConfigAsCodeTest.exportExportConfiguration
  and BasicConfigAsCodeTest.exportBasicConfiguration
* github-branch-source: GitHubAppCredentialsJCasCCompatibilityTest.should_support_configuration_export
* ldap: CascSecurityRealmTest.export_ldap_no_secret

All 3 of those plugins have been released with updated tests to
adapt to the improved behavior of Configuration as code plugin
1985.vdda_32d0c4ea_b_, but the azure-vm-agents and ldap releases both
require 2.516.x.  Exclude those tests until we drop plugin BOM support
for the 2.504.x line.

Detected in tests of 2.528.1-rc by @lemeurherve in pull request:

* https://github.com/jenkinsci/bom/pull/5770

Was not detected with last week's BOM release because in order to save
money we only test the weekly, oldest, and newest LTS versions when
performing a plugin BOM release.  Last week, 2.504.x was not included
in that test because we tested 2.492.x, 2.516.x, and weekly.  This week
we dropped the 2.492.x line and added the 2.528..x line, so we will test
2.504.x, 2.528.x, and weekly.