Test gitlab API 6.2.0 coordinated update
Rather than create a compatibility layer for the breaking changes between
GitLab API 5.8 and 6.x, we're planning a coordinated release of the
API plugin and the three plugins that depend on it.  Users will need to
upgrade the API plugin and the dependent plugins at the same time.

Tests the following 4 pull requests:

* https://github.com/jenkinsci/gitlab-api-plugin/pull/121
* https://github.com/jenkinsci/gitlab-branch-source-plugin/pull/517
* https://github.com/jenkinsci/gitlab-logo-plugin/pull/133
* https://github.com/jenkinsci/gitlab-oauth-plugin/pull/201

Testing done:

* Confirmed the 6.2.0 API upgrade works in each of the pull requests
* Confirmed this set of changes pass their own tests locally on the weekly line